requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
assemblyai==0.29.0
requests==2.31.0
google-generativeai==0.8.0
python-dotenv==1.0.0
pydantic==2.5.0
